+++ src/dependencies.js
@@ -1,6 +1,10 @@
 const dependencies = [
   {
     checked: true,
+    name: '@bobthered/tailwindcss-palette-generator',
+  },
+  {
+    checked: true,
     name: 'dotenv',
   },
   {
@@ -22,6 +26,26 @@ const dependencies = [
   {
     checked: true,
     name: 'less-middleware',
+  },
+  {
+    checked: true,
+    name: 'nodemon',
+  },
+  {
+    checked: true,
+    name: '@rollup/plugin-commonjs',
+  },
+  {
+    checked: true,
+    name: '@rollup/plugin-node-resolve',
+  },
+  {
+    checked: true,
+    name: 'rollup-plugin-terser',
+  },
+  {
+    checked: true,
+    name: 'tailwindcss',
   },
 ];